Basic create from CSV flow working. WIP overall
