+++ ClientApp/src/components/Home.tsx
@@ -52,7 +52,7 @@ const Home = (): JSX.Element => {
           }}
         >
           <IconButton
-            label="Exiting employees"
+            label="Active employees"
             iconName="user-check"
             colorType="outline-primary"
             marginClasses="mr-3"
