+++ ClientApp/src/helpers/labelHelper.tsx
@@ -27,9 +27,8 @@ const fieldLabels: { [key: string]: string } = {
   locationCity: 'Location city',
   originalHireDate: 'Original hire date',
   lastDayWorkedDate: 'Last day worked date',
-  effectiveDate: 'Effective date',
+  effectiveDate: 'Hire effective date',
   reason: 'Reason',
-  phone: 'Phone',
   appointmentStatus: 'Appointment status',
   positionCode: 'Position code',
   positionTitle: 'Position title',
