+++ ClientApp/src/types/EmployeeStatus.ts
@@ -3,10 +3,8 @@
 import { ISelectOption } from '../components/Employees/EditableSelect'
 
 export enum EmployeeStatusEnum {
-  Exiting = 'Exiting',
+  Active = 'Active',
   SurveyComplete = 'SurveyComplete',
-  SnailMailSent = 'SnailMailSent',
-  NotExiting = 'NotExiting',
   OutOfScope = 'OutOfScope',
   Declined = 'Declined',
   Expired = 'Expired'
@@ -35,29 +33,17 @@ export class EmployeeStatus {
     this.description = description
   }
 
-  static EXITING: EmployeeStatus = new EmployeeStatus(
-    EmployeeStatusEnum.Exiting,
+  static ACTIVE: EmployeeStatus = new EmployeeStatus(
+    EmployeeStatusEnum.Active,
     EmployeeStatusStateEnum.Active,
-    'Exiting',
-    'Employee is exiting.'
+    'Active',
+    'Employee is active.'
   )
   static SURVEY_COMPLETE: EmployeeStatus = new EmployeeStatus(
     EmployeeStatusEnum.SurveyComplete,
     EmployeeStatusStateEnum.Final,
     'Survey: Complete',
     'Survey has been finished.'
-  )
-  static SNAIL_MAIL_SENT: EmployeeStatus = new EmployeeStatus(
-    EmployeeStatusEnum.SnailMailSent,
-    EmployeeStatusStateEnum.Active,
-    'Snail mail sent',
-    'Snail mail has been sent.'
-  )
-  static NOT_EXITING: EmployeeStatus = new EmployeeStatus(
-    EmployeeStatusEnum.NotExiting,
-    EmployeeStatusStateEnum.Final,
-    'Employee not exiting',
-    'This employee is not actually exiting.'
   )
   static INELIGIBLE_OTHER: EmployeeStatus = new EmployeeStatus(
     EmployeeStatusEnum.OutOfScope,
@@ -79,10 +65,8 @@ export class EmployeeStatus {
   )
 
   static array = (): EmployeeStatus[] => [
-    EmployeeStatus.EXITING,
+    EmployeeStatus.ACTIVE,
     EmployeeStatus.SURVEY_COMPLETE,
-    EmployeeStatus.SNAIL_MAIL_SENT,
-    EmployeeStatus.NOT_EXITING,
     EmployeeStatus.INELIGIBLE_OTHER,
     EmployeeStatus.DECLINED,
     EmployeeStatus.EXPIRED
