+++ Controllers/Services/CallWeb/Models/CallWebPostDto.cs
@@ -6,37 +6,39 @@ namespace NewJobSurveyAdmin.Services.CallWeb
     {
         public string EmployeeId { get; set; }
         public string PreferredEmail { get; set; }
-        public string Ministry { get; set; }
-        public string PositionTitle { get; set; }
-        public string AppointmentStatus { get; set; }
-        public string PreferredFirstName { get; set; }
-        public string LastName { get; set; }
+        // public string Ministry { get; set; }
+        // public string PositionTitle { get; set; }
+        // public string AppointmentStatus { get; set; }
+        // public string PreferredFirstName { get; set; }
+        // public string LastName { get; set; }
         // public string LeaveReason { get; set; }
         // public string LeaveCode { get; set; }
-        // public string EffectiveDate { get; set; }
+        public string EffectiveDate { get; set; }
         // public string ExitCount { get; set; }
-        public string AdditionalJobCount { get; set; }
-        public string CurrentStatus { get; set; }
-        public string SurveyWindowFlag { get; set; }
+        // public string AdditionalJobCount { get; set; }
+        // public string CurrentStatus { get; set; }
+        // public string SurveyWindowFlag { get; set; }
 
         public static CallWebPostDto FromEmployee(Employee employee)
         {
+            // TODO: FIX THIS
             return new CallWebPostDto()
             {
                 EmployeeId = employee.GovernmentEmployeeId,
                 PreferredEmail = employee.PreferredEmail,
-                Ministry = employee.Organization,
-                PositionTitle = employee.PositionTitle,
-                AppointmentStatus = employee.AppointmentStatus,
-                PreferredFirstName = employee.PreferredFirstName,
-                LastName = employee.LastName,
+                EffectiveDate = employee.EffectiveDate.ToString("yyyy-MM-dd"),
+                // Ministry = employee.Organization,
+                // PositionTitle = employee.PositionTitle,
+                // AppointmentStatus = employee.AppointmentStatus,
+                // PreferredFirstName = employee.PreferredFirstName,
+                // LastName = employee.LastName,
                 // LeaveReason = employee.Reason,
                 // LeaveCode = employee.LeaveCode,
                 // EffectiveDate = employee.EffectiveDate.ToString("yyyy-MM-dd"),
                 // ExitCount = employee.ExitCount,
-                AdditionalJobCount = employee.RecordCount,
-                CurrentStatus = employee.CurrentEmployeeStatusCode,
-                SurveyWindowFlag = employee.SurveyWindowFlag()
+                // AdditionalJobCount = employee.RecordCount,
+                // CurrentStatus = employee.CurrentEmployeeStatusCode,
+                // SurveyWindowFlag = employee.SurveyWindowFlag()
             };
         }
     }
